fix margin when drawing the line numbers
This was never really an oxygen issue. Instead, for instance the font
"Courier", a digit has a real character width. Calculating with int
sums off and reflects in this bad offset.

FIXED-IN: 4.11.2
BUG: 312321

+++ part/view/kateviewhelpers.h
@@ -217,7 +217,7 @@ class KateIconBorder : public QWidget
 
     int m_cachedLNWidth;
 
-    int m_maxCharWidth;
+    qreal m_maxCharWidth;
     int iconPaneWidth;
     int m_annotationBorderWidth;
